+++ src/components/common/Button.jsx
@@ -2,6 +2,7 @@ import { useContext } from "react";
 import { IoCheckmarkOutline } from "react-icons/io5";
 import styled from "styled-components";
 import { SignUpContext } from "../context/SignUpContext";
+import { useNavigate } from "react-router-dom";
 export default function Button(props) {
   const {
     agreement,
@@ -11,6 +12,8 @@ export default function Button(props) {
     signUpStep,
     overlapCheck,
   } = useContext(SignUpContext);
+
+  const navigate = useNavigate();
 
   return (
     <>
@@ -34,7 +37,7 @@ export default function Button(props) {
             });
           }
           if (props.text == "next" && agreement == 3) {
-            setSignUpStep("CreateName");
+            navigate("/CreateName");
           }
         }}
         disabled={
