+++ src/router.js
@@ -1,13 +1,18 @@
 import { createBrowserRouter } from "react-router-dom";
-import SignUp from "./pages/SignUp";
 import App from "./pages/App";
 import IsLoginTemplate from "./pages/IsLoginTemplate";
 import Login from "./pages/Login";
+import CreateName from "./pages/CreateName";
+import CollectPersonalData from "./components/collectPersonalData/CollectPersonalData";
 
 const router = createBrowserRouter([
   {
     path: "/SignUp",
-    element: <SignUp />,
+    element: <CollectPersonalData />,
+  },
+  {
+    path: "/CreateName",
+    element: <CreateName />,
   },
   {
     path: "/Login",
